# Conflicts:
#	src/main/java/pro/sky/telegrambot/repository/PetRepository.java
#	src/main/java/pro/sky/telegrambot/service/PetService.java

+++ src/main/java/pro/sky/telegrambot/repository/PhotoRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface PhotoRepository extends JpaRepository<Photo, Long> {
 
-    Optional<Photo> findPhotoById(Long Id);
+    Optional<Photo> findPhotoById(int Id);
 }
